requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Jinja2==3.0.3
SQLAlchemy==1.4.27
Werkzeug==2.0.2
WTForms==3.0.0
click==8.0.3
itsdangerous==2.0.1
MarkupSafe==2.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.9
Flask-Migrate==4.0.7
python-dotenv
Pillow==10.4.0